[fixed] /moveFurniture command increments that was broken in an earlier revision.
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk@3224 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ MMOCoreORB/src/server/zone/objects/creature/commands/MoveFurnitureCommand.h
@@ -146,7 +146,7 @@ public:
 		}
 
 		String dir;
-		int dist = 0;
+		float dist = 0.f;
 
 		try {
 			UnicodeTokenizer tokenizer(arguments.toString());
@@ -163,7 +163,7 @@ public:
 			return INVALIDPARAMETERS;
 		}
 
-		if (dist < 1 || dist > 500) {
+		if (dist < 1.f || dist > 500.f) {
 			creature->sendSystemMessage("@player_structure:movefurniture_params"); //The amount to move must be between 1 and 500.
 			return INVALIDPARAMETERS;
 		}
